+++ OwlOscUnity/Assets/OwlOSC/Runtime/UnitySystemConsoleRedirector.cs
@@ -1,0 +1,76 @@
+using System;
+using System.IO;
+using System.Text;
+using UnityEngine;
+ 
+/// <summary>
+/// Redirects writes to System.Console to Unity3D's Debug.Log.
+/// </summary>
+/// <author>
+/// Jackson Dunstan, http://jacksondunstan.com/articles/2986
+/// </author>
+/// 
+namespace OWL{
+
+	public static class UnitySystemConsoleRedirector
+	{
+		private class UnityTextWriter : TextWriter
+		{
+			private StringBuilder buffer = new StringBuilder();
+	
+			public override void Flush()
+			{
+				Debug.Log(buffer.ToString());
+				buffer.Length = 0;
+			}
+	
+			public override void Write(string value)
+			{
+				buffer.Append(value);
+				if (value != null)
+				{
+					var len = value.Length;
+					if (len > 0)
+					{
+						var lastChar = value [len - 1];
+						if (lastChar == '\n')
+						{
+							Flush();
+						}
+					}
+				}
+			}
+	
+			public override void Write(char value)
+			{
+				buffer.Append(value);
+				if (value == '\n')
+				{
+					Flush();
+				}
+			}
+	
+			public override void Write(char[] value, int index, int count)
+			{
+				Write(new string (value, index, count));
+			}
+	
+			public override Encoding Encoding
+			{
+				get { return Encoding.Default; }
+			}
+		}
+	
+		/// <summary>
+		/// Set Redirect to Unity Console
+		/// </summary>
+		/// <param name="redirect">If True redirect, if False set to default</param>
+		public static void Redirect(bool redirect = true)
+		{
+			if(redirect)
+				Console.SetOut(new UnityTextWriter());
+			else
+				Console.SetOut(new StreamWriter(Console.OpenStandardOutput()));
+		}
+	}
+}